rest-interface: delete file, upload file, append file size into list file.

+++ src/main/java/com/github/johnsonmoon/java/ftp/server/controller/FtpController.java
@@ -3,6 +3,8 @@ package com.github.johnsonmoon.java.ftp.server.controller;
 import com.github.johnsonmoon.java.ftp.server.common.ext.FtpException;
 import com.github.johnsonmoon.java.ftp.server.entity.FileDefine;
 import com.github.johnsonmoon.java.ftp.server.entity.FileDefineTypeEnum;
+import com.github.johnsonmoon.java.ftp.server.entity.param.UploadFileParam;
+import com.sun.org.apache.xpath.internal.operations.Bool;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.slf4j.Logger;
@@ -49,10 +51,70 @@ public class FtpController {
                 if (f.isFile()) {
                     fileDefine.setType(FileDefineTypeEnum.FILE.getType());
                 }
+                fileDefine.setSize(f.length());
                 fileDefines.add(fileDefine);
             }
         }
         return fileDefines;
+    }
+
+    @ApiOperation("Upload file.")
+    @PostMapping(value = "/upload", consumes = "multipart/form-data")
+    public Boolean uploadFile(UploadFileParam uploadFileParam) {
+        if (uploadFileParam.getName() == null || uploadFileParam.getName().isEmpty()
+                || uploadFileParam.getTargetDir() == null || uploadFileParam.getTargetDir().isEmpty()
+                || uploadFileParam.getFile() == null) {
+            throw new FtpException("Name or targetDir or file is null!");
+        }
+        File dir = new File(uploadFileParam.getTargetDir());
+        if (!dir.exists()) {
+            throw new FtpException("TargetDir not exists!");
+        }
+        if (!dir.isDirectory()) {
+            throw new FtpException(String.format("%s is not a directory!", dir));
+        }
+        String filePathName = uploadFileParam.getTargetDir() + File.separator + uploadFileParam.getName();
+        File newFile = new File(filePathName);
+        if (newFile.exists()) {
+            if (!newFile.delete()) {
+                throw new FtpException("Old file in target directory exist with same fileName and could not be deleted.");
+            }
+        }
+        try {
+            if (!newFile.createNewFile()) {
+                throw new FtpException("Create new file failed.");
+            }
+        } catch (Exception e) {
+            logger.warn(e.getMessage(), e);
+            throw new FtpException("Create new file failed.");
+        }
+        InputStream inputStream = null;
+        FileOutputStream outputStream = null;
+        try {
+            inputStream = uploadFileParam.getFile().getInputStream();
+            outputStream = new FileOutputStream(newFile);
+            byte[] bytes = new byte[32];
+            int length;
+            while ((length = inputStream.read(bytes)) > 0) {
+                outputStream.write(bytes, 0, length);
+            }
+            outputStream.flush();
+        } catch (Exception e) {
+            logger.warn(e.getMessage(), e);
+            throw new FtpException("Upload file failed.");
+        } finally {
+            try {
+                if (inputStream != null) {
+                    inputStream.close();
+                }
+                if (outputStream != null) {
+                    outputStream.close();
+                }
+            } catch (Exception e) {
+                logger.warn(e.getMessage(), e);
+            }
+        }
+        return true;
     }
 
     @ApiOperation("Download file.")
@@ -150,6 +212,22 @@ public class FtpController {
             } catch (Exception e) {
                 logger.warn(e.getMessage(), e);
             }
+        }
+        return f.delete();
+    }
+
+    @ApiOperation("Delete file.")
+    @GetMapping("/delete")
+    public Boolean deleteFile(@RequestParam("file") String file) {
+        if (file == null || file.isEmpty()) {
+            throw new FtpException("file is null!");
+        }
+        File f = new File(file);
+        if (!f.exists()) {
+            throw new FtpException("file not exist.");
+        }
+        if (!f.isFile()) {
+            throw new FtpException("file is not a file.");
         }
         return f.delete();
     }
